Rename component files as per convention.
