+++ airflow/ui/src/components/ui/index.ts
@@ -32,3 +32,5 @@ export * from "./ProgressBar";
 export * from "./Menu";
 export * from "./Accordion";
 export * from "./Status";
+export * from "./Button";
+export * from "./Toaster";
